feat: add WorkspaceGroupUserDeleteModal

fix: fix


fix: fix minor things
