+++ packages/language-pack/en.json
@@ -2843,6 +2843,7 @@
 				"ALT_S_ADD_WORKSPACE": "Workspaces successfully added",
 				"ALT_S_CREATE_WORKSPACE": "Workspace group successfully created",
 				"ALT_S_EDIT_WORKSPACE": "Workspace Group successfully updated",
+				"ASSOCIATED_WORKSPACES": "Associated Workspaces",
 				"CREATE_DROP_DOWN_PLACEHOLDER": "Select Workspaces",
 				"CREATE_GROUP_NAME": "Name",
 				"CREATE_GROUP_NAME_PLACEHOLDER": "Enter Name",
@@ -2858,6 +2859,8 @@
 				"EDIT_GROUP_NAME": "Name",
 				"EDIT_GROUP_NAME_PLACEHOLDER": "Enter Name",
 				"EDIT_TITLE": "Edit Workspace Group",
+				"REMOVE_GROUP_USER_DESCRIPTION": "The selected user(s) will be removed from the workspace group, as well as from all workspaces contained within it.",
+				"REMOVE_GROUP_USER_TITLE": "Selected User",
 				"ROLE_DROP_DOWN_PLACEHOLDER": "Select a role",
 				"ROLE_SELECT_DROP_DOWN_TITLE": "Role",
 				"USER_DROP_DOWN_PLACEHOLDER": "Add Users",
