+++ packages/language-pack/ja.json
@@ -2843,6 +2843,7 @@
 				"ALT_S_ADD_WORKSPACE": "",
 				"ALT_S_CREATE_WORKSPACE": "ワークスペースグループの作成が完了しました",
 				"ALT_S_EDIT_WORKSPACE": "ワークスペースグループが正常に更新されました",
+				"ASSOCIATED_WORKSPACES": "",
 				"CREATE_DROP_DOWN_PLACEHOLDER": "ワークスペースの選択",
 				"CREATE_GROUP_NAME": "名前",
 				"CREATE_GROUP_NAME_PLACEHOLDER": "名前を入力",
@@ -2858,6 +2859,8 @@
 				"EDIT_GROUP_NAME": "名前",
 				"EDIT_GROUP_NAME_PLACEHOLDER": "名前を入力",
 				"EDIT_TITLE": "ワークスペースグループを編集",
+				"REMOVE_GROUP_USER_DESCRIPTION": "",
+				"REMOVE_GROUP_USER_TITLE": "",
 				"ROLE_DROP_DOWN_PLACEHOLDER": "役割を選択してください",
 				"ROLE_SELECT_DROP_DOWN_TITLE": "ロール",
 				"USER_DROP_DOWN_PLACEHOLDER": "ユーザー追加",
