+++ packages/language-pack/ko.json
@@ -2843,6 +2843,7 @@
 				"ALT_S_ADD_WORKSPACE": "워크스페이스 추가 완료",
 				"ALT_S_CREATE_WORKSPACE": "워크스페이스 그룹 생성 완료",
 				"ALT_S_EDIT_WORKSPACE": "워크스페이스 그룹 변경 완료",
+				"ASSOCIATED_WORKSPACES": "",
 				"CREATE_DROP_DOWN_PLACEHOLDER": "워크스페이스 선택",
 				"CREATE_GROUP_NAME": "이름",
 				"CREATE_GROUP_NAME_PLACEHOLDER": "이름 입력",
@@ -2858,6 +2859,8 @@
 				"EDIT_GROUP_NAME": "이름",
 				"EDIT_GROUP_NAME_PLACEHOLDER": "이름 입력",
 				"EDIT_TITLE": "워크스페이스 그룹 편집",
+				"REMOVE_GROUP_USER_DESCRIPTION": "",
+				"REMOVE_GROUP_USER_TITLE": "",
 				"ROLE_DROP_DOWN_PLACEHOLDER": "역할(Role) 선택",
 				"ROLE_SELECT_DROP_DOWN_TITLE": "역할(Role)",
 				"USER_DROP_DOWN_PLACEHOLDER": "사용자 추가",
